Move ledger example to an integration test
